third_party/WebKit/LayoutTests/crypto/subtle/ecdh/deriveKey-hmac-expected.txt

Tests deriveKey() using ECDH to make HMAC keys

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing the private key...

Importing the public key...

Deriving an HMAC 136 bit key...

PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is true
PASS key.algorithm.name is "HMAC"
PASS key.algorithm.hash.name is "SHA-1"
PASS key.algorithm.length is 136
PASS key.usages.join(',') is "sign"
PASS: Derived Bytes should be [0117d54d84379d0fd385be068455a77a53] and was
Deriving an HMAC 256 bit key...

PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is true
PASS key.algorithm.name is "HMAC"
PASS key.algorithm.hash.name is "SHA-256"
PASS key.algorithm.length is 512
PASS key.usages.join(',') is "sign,verify"
PASS: Derived Bytes should be [0117d54d84379d0fd385be068455a77a5366ab534ff172ab0a121f37d180dccd19607abb0c41cb9f6f12b01303ac4a69dc2d1d05180181fd496d9769b46bffec] and was
Deriving an HMAC 256 bit key...

PASS key.type is "secret"
PASS typeof key.extractable is 'boolean'
PASS key.extractable is false
PASS key.algorithm.name is "HMAC"
PASS key.algorithm.hash.name is "SHA-256"
PASS key.algorithm.length is 256
PASS key.usages.join(',') is "verify"
PASS successfullyParsed is true

TEST COMPLETE
